+++ packages/flui-spec-js/src/types/index.ts
@@ -2,8 +2,13 @@ export * from './api-version';
 export * from './enums';
 export * from './catalog-app';
 export * from './application';
+export * from './access-policy';
 
 import type { CatalogAppManifest } from './catalog-app';
 import type { ApplicationManifest } from './application';
+import type { AccessPolicyManifest } from './access-policy';
 
-export type FluiManifest = CatalogAppManifest | ApplicationManifest;
+export type FluiManifest =
+  | CatalogAppManifest
+  | ApplicationManifest
+  | AccessPolicyManifest;